+++ CHANGELOG.md
@@ -13,9 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Implemented a new task scheduler which resolves several long-standing issues:
+
   - Issues with yielding across the c-call/metamethod boundary no longer occur when calling certain async APIs that Lune provides.
   - Ordering of interleaved calls to `task.spawn/task.defer` is now completely deterministic, defer is now guaranteed to run last even in these cases.
   - The minimum wait time possible when using `task.wait` and minimum delay time using `task.delay` are now much smaller, and only limited by the underlying OS implementation. For most systems this means `task.wait` and `task.delay` are now accurate down to about 5 milliseconds or less.
+
+- Added a new function `fs.move` to move / rename a file or directory from one path to another.
 
 ### Changed